Add showModal to delete method

+++ pages/lessons/show.js
@@ -29,16 +29,28 @@ Page({
 
   deleteConfirmed() {
     const page = this;
-    wx.request({
-      url: `${app.globalData.baseUrl}/lessons/${page.data.id}`,
-      method: "DELETE",
-      header: app.globalData.header,
+    wx.showModal({
+      title: 'Are you sure?',
+      confirmText: 'Yes',
+      confirmColor: 'green',
+      cancelText: 'No',
+      cancelColor: 'red',
       success(res) {
-        console.log(res);
-        wx.navigateBack();
+        if (res.confirm) {
+          wx.request({
+            url: `${app.globalData.baseUrl}/lessons/${page.data.id}`,
+            method: "DELETE",
+            header: app.globalData.header,
+            success(res) {
+              // console.log(res);
+              wx.navigateBack();
+            }
+          })
+        }
       }
     })
   },
+
   /**
    * Lifecycle function--Called when page load
    */
